+++ Problems/Day05B.cs
@@ -2,7 +2,7 @@
 
 namespace Advent_of_Code_2023;
 
-public partial class Day05B : Problem<Day05B.Input, uint> {
+public partial class Day05B : Problem<Day05B.Input, long> {
     public readonly record struct Input(Input.Range[] Seeds, Input.Mapping[] Mappings) {
         public readonly struct Mapping {
             public Mapping(IEnumerable<RangeMap> rangeMaps) {
@@ -12,7 +12,7 @@ public partial class Day05B : Problem<Day05B.Input, uint> {
             public IEnumerable<Range> Map(in Range range) {
                 List<Range> ranges = new();
 
-                uint mapEnd = range.Start;
+                long mapEnd = range.Start;
                 foreach (RangeMap rangeMap in RangeMaps) {
                     Range fromIntersect = rangeMap.From.Intersect(range);
 
@@ -34,24 +34,24 @@ public partial class Day05B : Problem<Day05B.Input, uint> {
             private RangeMap[] RangeMaps { get; }
         }
 
-        public readonly record struct Range(uint Start, uint Count) {
-            public uint End => Start + Count;
+        public readonly record struct Range(long Start, long Count) {
+            public long End => Start + Count;
 
             public bool Empty => Count == 0;
 
             public Range Intersect(in Range that) {
-                uint start = Math.Max(Start, that.Start);
-                uint end   = Math.Min(End, that.End);
+                long start = Math.Max(Start, that.Start);
+                long end   = Math.Min(End, that.End);
                 if (end <= start) return new Range(Start, 0);
                 return new Range(start, end - start);
             }
         }
 
-        public readonly record struct RangeMap(uint ToStart, uint FromStart, uint Count) {
+        public readonly record struct RangeMap(long ToStart, long FromStart, long Count) {
             public Range From => new(FromStart, Count);
             public Range To   => new(ToStart, Count);
 
-            public uint Map(uint value) => ToStart + (value - FromStart);
+            public long Map(long value) => ToStart + (value - FromStart);
 
             public Range Map(in Range range) => range with { Start = Map(range.Start) };
         }
@@ -70,7 +70,7 @@ public partial class Day05B : Problem<Day05B.Input, uint> {
 
         Input.Range[] seeds;
         {
-            uint[] numbers = Numbers(blocks[0]);
+            long[] numbers = Numbers(blocks[0]);
             seeds = new Input.Range[numbers.Length / 2];
             for (int i = 0; i < seeds.Length; i++) {
                 seeds[i] = new Input.Range(numbers[2 * i + 0], numbers[2 * i + 1]);
@@ -90,10 +90,10 @@ public partial class Day05B : Problem<Day05B.Input, uint> {
                .ToArray()
         );
 
-        static uint[] Numbers(string s) => NumberRegex().Matches(s).Select(m => uint.Parse(m.Value)).ToArray();
+        static long[] Numbers(string s) => NumberRegex().Matches(s).Select(m => long.Parse(m.Value)).ToArray();
     }
 
-    protected override uint Solve(Input input) {
+    protected override long Solve(Input input) {
         return input.Map(input.Seeds).Min(range => range.Start);
     }
